backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-multiselectfield==0.1.12
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
